editor/libeditor/tests/test_bug1181130-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1181130
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1181130</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1181130">Mozilla Bug 1181130</a>
<p id="display"></p>
<div id="container" contenteditable="true">
  editable div
  <div id="noneditable" contenteditable="false">
    non-editable div
  </div>
</div>
<script type="application/javascript">
/** Test for Bug 1181130 **/
var container = document.getElementById("container");
var noneditable = document.getElementById("noneditable");

SimpleTest.waitForExplicitFinish();

SimpleTest.waitForFocus(function() {
  var nonHTMLElement = document.createElementNS("http://www.example.com", "element");
  nonHTMLElement.innerHTML = '<div contenteditable="true">nested editable div</div>';
  noneditable.appendChild(nonHTMLElement);

  synthesizeMouseAtCenter(noneditable, {});
  ok(!document.getSelection().toString().includes("nested editable div"),
     "Selection should not include non-editable content");

  SimpleTest.finish();
});
</script>
<pre id="test">
</pre>
</body>
</html>
